Added navigation to recipeScreen

+++ app/src/main/java/com/example/recipefeed/data/model/Recipe.kt
@@ -1,7 +1,7 @@
 package com.example.recipefeed.data.model
 
 data class Recipe(
-    val id: Long = 1L,
+    val id: Int = 1,
     val recipeName: String = "testRecipeName",
     val ingredients: String = "testIngredients",
     val description: String = "testDescription",
